数据类型

+++ modules.py
@@ -48,7 +48,7 @@ def get_token_embeddings(vocab_size, num_units, embedding_name, hp, zero_pad=Tru
 
     if hp.pretrain==1:
         with open(os.path.join(hp.pretrained_emb_path, '%s.emb' %(embedding_name)), 'rb') as f:
-            emb_value = pickle.load(f)
+            emb_value = pickle.load(f).astype(np.float32)
 
         print('(',vocab_size, num_units,')')
         print(emb_value.shape)
